+++ src/main/java/duke/task/TodoTask.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class TodoTask extends Task {
     public TodoTask(String description) {
         super(description);
